gateway/requirements.txt
attrs==23.1.0
click==7.1.2
Flask==1.1.4
grpcio-tools~=1.30
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
more-itertools==10.1.0
packaging==23.1
pluggy==0.13.1
py==1.11.0
pytest==5.4.3
wcwidth==0.2.6
Werkzeug==1.0.1
grpc-interceptor ~= 0.12.0
